requirements.txt
ultralytics
scikit-learn
Pillow
openimages
